disable skopei / go-about api's update open-bikelocker remove twilio support
