+++ app/.meteor/packages
@@ -6,7 +6,7 @@
 
 meteor-base@1.1.0             # Packages every Meteor app needs to have
 mobile-experience@1.0.5       # Packages for a great mobile UX
-mongo@1.2.2                  # The database Meteor supports right now
+mongo@1.2.3                  # The database Meteor supports right now
 blaze-html-templates@1.0.4    # Compile .html files into Meteor Blaze views
 reactive-var@1.0.11           # Reactive variable for tracker
 jquery@1.11.10                # Helpful client-side library
