+++ app/.meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.5.2.2
+METEOR@1.5.4.1
