+++ app/.meteor/versions
@@ -1,9 +1,9 @@
-accounts-base@1.3.4
+accounts-base@1.3.6
 accounts-facebook@1.2.1
 accounts-github@1.3.0
 accounts-google@1.2.0
 accounts-oauth@1.1.15
-accounts-password@1.4.0
+accounts-password@1.4.2
 accounts-twitter@1.3.0
 accounts-ui@1.1.9
 accounts-ui-unstyled@1.2.1
@@ -39,13 +39,13 @@ ecmascript-runtime-server@0.4.1
 ejson@1.0.14
 email@1.2.3
 es5-shim@4.6.15
-facebook-config-ui@1.0.0
+facebook-config-ui@1.0.1
 facebook-oauth@1.3.2
 geojson-utils@1.0.10
 github-config-ui@1.0.0
 github-oauth@1.2.0
 google-config-ui@1.0.0
-google-oauth@1.2.4
+google-oauth@1.2.5
 hot-code-push@1.0.4
 html-tools@1.0.11
 htmljs@1.0.11
@@ -58,24 +58,24 @@ launch-screen@1.1.1
 less@2.7.9
 livedata@1.0.18
 localstorage@1.1.1
-logging@1.1.17
+logging@1.1.19
 mdg:validation-error@0.5.1
 meteor@1.7.2
 meteor-base@1.1.0
 meteortoys:toykit@4.0.1
 minifier-css@1.2.16
 minifier-js@2.1.4
-minimongo@1.3.2
+minimongo@1.3.3
 mobile-experience@1.0.5
 mobile-status-bar@1.0.14
 modules@0.10.0
 modules-runtime@0.8.0
-mongo@1.2.2
+mongo@1.2.3
 mongo-dev-server@1.0.1
 mongo-id@1.0.6
 msavin:mongol@4.0.1
 npm-bcrypt@0.9.3
-npm-mongo@2.2.30
+npm-mongo@2.2.33
 oauth@1.1.13
 oauth1@1.1.11
 oauth2@1.1.11
