+++ app/imports/client/commonbike-ui.jsx
@@ -40,12 +40,6 @@ class CommonBikeUI extends Component {
     })
   }
 
-  onTestGoAbout() {
-    Meteor.call('goabout.checklocations', (error, result) => {
-      console.log(error || result)
-    })
-  }
-
   onTestSkopei() {
     Meteor.call('skopei.checklocations', (error, result) => {
       console.log(error || result)
@@ -74,9 +68,6 @@ class CommonBikeUI extends Component {
 
         <h2>Skopei</h2>
         <RaisedButton onClick={this.onTestSkopei}>Test Skopei API</RaisedButton>
-
-        <h2>GoAbout</h2>
-        <RaisedButton onClick={this.onTestGoAbout}>Test GoAbout API</RaisedButton>
 
         <h2>Button</h2>
         <Button>Klik hier! ~ Button</Button>
